✅ (dashboard): Add audit sandbox-summary query tests
